+++ 3D Graphics Engine/src/Node.cpp
@@ -56,7 +56,7 @@ int Node::getNumberOfChildren()
 
 bool Node::seekNonExistentChildren()
 {
-    for (int i = 0; i < children.size(); i++)
+    for (size_t i = 0; i < children.size(); i++)
     {
         if (children.at(i) == NULL)
         {
